Commands for viewing the documentation
Added commands for viewing the generated documentation in the default
browser
