+++ python/sphinx_docs/pyproject.toml
@@ -18,6 +18,7 @@ sphinx-versioned-docs = "^1.3.1"
 
 [tool.poetry.scripts]
 opmdoc-download-files = "opm_python_docs.download_files:main"
+opmdoc-view-doc = "opm_python_docs.view_docs:main"
 
 [build-system]
 requires = ["poetry-core"]
